Adopted src structure. Added setup.py, LICENSE, README.rst. Found a critical bug in model._linspace(): round(window / resolution) returns a float, and seems that depending on numpy version this may or may not be allowed for the linspace num argument. int(round(...)) fixes.

+++ src/qt_nmr/model/model.py
@@ -36,7 +36,7 @@ class Model(QObject):
         max_ = peaklist[-1][0]
         lin_min, lin_max = min_ - margin, max_ + margin
         window = lin_max - lin_min  # Hz
-        datapoints = round(window / resolution)
+        datapoints = int(round(window / resolution))
         logger.debug(f'making linspace for {min_} {max_} {resolution}')
         return np.linspace(lin_min, lin_max, datapoints)
